data classを移動
