--- app/src/main/java/com/sereno/serenoandroidapp/Clouds.kt
@@ -1,5 +1,0 @@
-package com.sereno.serenoandroidapp
-
-data class Clouds(
-    val all: Int
-)
