--- app/src/main/java/com/sereno/serenoandroidapp/Coord.kt
@@ -1,6 +1,0 @@
-package com.sereno.serenoandroidapp
-
-data class Coord(
-    val lat: Double,
-    val lon: Double
-)
