--- app/src/main/java/com/sereno/serenoandroidapp/Wind.kt
@@ -1,6 +1,0 @@
-package com.sereno.serenoandroidapp
-
-data class Wind(
-    val deg: Int,
-    val speed: Double
-)
