+++ app/src/main/java/com/sereno/serenoandroidapp/models/apis/openweathermap/currentweatherdata/Clouds.kt
@@ -1,0 +1,5 @@
+package com.sereno.serenoandroidapp.models.apis.openweathermap.currentweatherdata
+
+data class Clouds(
+    val all: Int
+)
