+++ app/src/main/java/com/sereno/serenoandroidapp/models/apis/openweathermap/currentweatherdata/Coord.kt
@@ -1,0 +1,6 @@
+package com.sereno.serenoandroidapp.models.apis.openweathermap.currentweatherdata
+
+data class Coord(
+    val lat: Double,
+    val lon: Double
+)
