+++ app/src/main/java/com/sereno/serenoandroidapp/models/apis/openweathermap/currentweatherdata/Main.kt
@@ -1,4 +1,4 @@
-package com.sereno.serenoandroidapp
+package com.sereno.serenoandroidapp.models.apis.openweathermap.currentweatherdata
 
 data class Main(
     val feels_like: Double,
