+++ app/src/main/java/com/sereno/serenoandroidapp/models/apis/openweathermap/currentweatherdata/Wind.kt
@@ -1,0 +1,6 @@
+package com.sereno.serenoandroidapp.models.apis.openweathermap.currentweatherdata
+
+data class Wind(
+    val deg: Int,
+    val speed: Double
+)
